NXP-28865: move createDatabase from JDBCMapper to RepositoryImpl
