+++ modules/core/nuxeo-core-storage-sql/nuxeo-core-storage-sql/src/main/java/org/nuxeo/ecm/core/storage/sql/SoftRefCachingMapper.java
@@ -83,11 +83,6 @@ public class SoftRefCachingMapper extends SoftRefCachingRowMapper implements Cac
     }
 
     @Override
-    public void createDatabase(String ddlMode) {
-        mapper.createDatabase(ddlMode);
-    }
-
-    @Override
     public Serializable getRootId(String repositoryId) {
         return mapper.getRootId(repositoryId);
     }
