+++ modules/core/nuxeo-core-storage-sql/nuxeo-core-storage-sql/src/main/java/org/nuxeo/ecm/core/storage/sql/UnifiedCachingMapper.java
@@ -84,11 +84,6 @@ public class UnifiedCachingMapper extends UnifiedCachingRowMapper implements Cac
     }
 
     @Override
-    public void createDatabase(String ddlMode) {
-        mapper.createDatabase(ddlMode);
-    }
-
-    @Override
     public Serializable getRootId(String repositoryId) {
         return mapper.getRootId(repositoryId);
     }
